Move ExternalAgentParsing to internal.utils
