+++ neuro_san/internals/graph/activations/calling_activation.py
@@ -27,13 +27,13 @@ from leaf_common.config.dictionary_overlay import DictionaryOverlay
 from neuro_san.internals.graph.activations.abstract_callable_activation import AbstractCallableActivation
 from neuro_san.internals.interfaces.agent_tool_factory import AgentToolFactory
 from neuro_san.internals.interfaces.callable_activation import CallableActivation
-from neuro_san.internals.journals.external_agent_parsing import ExternalAgentParsing
 from neuro_san.internals.journals.journal import Journal
 from neuro_san.internals.run_context.factory.run_context_factory import RunContextFactory
 from neuro_san.internals.run_context.interfaces.run import Run
 from neuro_san.internals.run_context.interfaces.run_context import RunContext
 from neuro_san.internals.run_context.interfaces.tool_call import ToolCall
 from neuro_san.internals.run_context.interfaces.tool_caller import ToolCaller
+from neuro_san.internals.utils.external_agent_parsing import ExternalAgentParsing
 
 
 class CallingActivation(AbstractCallableActivation, ToolCaller):
